src/portfolio/dbt_requirements.txt
dbt-core==1.8.6
dbt-snowflake==1.8.3
sqlfluff==3.2.0
